% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree_score.R
\name{agree_score}
\alias{agree_score}
\title{Agreement Scores}
\usage{
agree_score(X)
}
\arguments{
\item{X}{an \eqn{n\times p}{nxp} matrix of votes, which can include \code{NA}.}
}
\value{
A data frame containing agreement scores and number of joint votes
for every pair of voters.
}
\description{
Calculate agreement scores among voters/legislators.
}
