% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pres.r
\name{calc_pres}
\alias{calc_pres}
\title{Calculate PRE}
\usage{
calc_pres(votes, initlv, glm_method = c("glm", "firth"), ...)
}
\arguments{
\item{votes}{A matrix of votes}

\item{initlv}{Initial estimate of latent variables}

\item{glm_method}{Use either binomial glm or firth logistic regression to calculate coefficients and PRE.}

\item{...}{Other arguments to be passed down}
}
\value{
A list with a matrix of PRE values and matrix of coefficients
from the various GLMs estimated to calculate PRE.
}
\description{
Calculates PRE for initialized latent variables.
}
\details{
Calculates the PRE for each indicator.  This is used to identify
which indicator is best predicted by which latent variable.
}
