% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best.R
\name{choose_best}
\alias{choose_best}
\title{Choose Best Dimension}
\usage{
choose_best(
  x,
  bestmin = 0,
  othermax = 0,
  ndim = NULL,
  choice_method = c("pre", "b"),
  ...
)
}
\arguments{
\item{x}{Depending on \code{choice_method} it could be a #Votes x #Dimensions 
matrix of PRE values or p-values.}

\item{bestmin}{Scalar indicating the minimum PRE value to be
considered "well predicted" bythe model.}

\item{othermax}{Scalar indicating the maximum PRE of the inferior dimension.
If \code{NULL}, the algorithm just chooses the first highest PRE, regardless
of the other PRE values.}

\item{ndim}{Scalar giving the number of dimensions to use.  The algorithm
will use the first \code{ndim} columns of \code{x} in the calculation.}

\item{choice_method}{String giving the method used to choose the best dimension.
By default, (the \code{"pre"} option) the best dimension is chosen by comparing 
PREs.  The \code{"b"} option chooses dimensions based on the significance of the 
coefficient relating the initialized latent variable to all of the votes. 
If \code{NULL}, all dimensions are used.}

\item{...}{Other arguments to be passed down - currently not implemented.}
}
\value{
A vector identifying which dimension was best or \code{NA}
if no dimension met both criteria.
}
\description{
Chooses latent dimension that best predicts the outcome.
}
\details{
The algorithm chooses the best dimension by identifying for each vote the
situation where the pre for variable j is greater than \code{bestmin} and
the pre for all other variables not equal to j is less than \code{othermin}.
}
