% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.r
\name{flip}
\alias{flip}
\title{Reflect Latent Variable Estimates}
\usage{
flip(data, id = "name", vars, time, ...)
}
\arguments{
\item{data}{A data frame or tibble produced with \code{gather_data}.}

\item{id}{Character string giving the name(s) of the ID variable(s) 
in the data.}

\item{vars}{Variables to be evaluated for reflection.  Generally, 
these would be all of the \code{DimX} variables.}

\item{time}{Character string giving the time variable in the 
data set.}

\item{...}{Other arguments to be passed down, currently not implemented.}
}
\value{
A data frame with time and a flip indicator for each 
evaluated dimension.  This can be merged with the original data and
can be used to make the appropriate reflections of the latent variables.
}
\description{
Identifies reflections that minimize the distance of  adjacent 
time-points on the same latent variable
}
