% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_data.r
\name{gather_data}
\alias{gather_data}
\title{Collect Data from Estimated Models}
\usage{
gather_data(
  x,
  ...,
  boot_sd = FALSE,
  Nboot = 100,
  dynIRT_control = list(threads = 1, verbose = TRUE, thresh = 1e-06, maxit = 500),
  orthogonalize = c("none", "gs", "pca")
)
}
\arguments{
\item{x}{An object of class \code{legR}.}

\item{...}{Other arguments to be passed down}

\item{boot_sd}{Logical indicating whether bootstrapped posterior standard deviations should be produced.}

\item{Nboot}{If \code{boot_sd = TRUE}, then how many bootstrap samples should be used to calculate the
posterior standard deviation.}

\item{dynIRT_control}{A list of control parameters for the dynamic IRT models.}

\item{orthogonalize}{Character string giving the kind of orthogonalization to be done,
including \code{"none"}, \code{"gs"} for Gram-Schmidt (i.e., residualization)
or \code{"pca"} indicating a principal components orthogonalization.  Both orthogonalization options
are done on the posterior means and that same transform is applied to all iterations to get
appropriate uncertainty estimates for the orthogonalized variables.}
}
\value{
A data frame with latent variable estimates.
}
\description{
Collects data from model and transforms it from wide- to long-form.  Also estimates bootstrapped posterior
standard deviations for the latent point estimates.
}
