% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPRE.r
\name{getPRE}
\alias{getPRE}
\title{Proportional and Expected Proportional Reductions in Error}
\usage{
getPRE(mod1)
}
\arguments{
\item{mod1}{A model of class \code{glm} (with family \code{binomial}),
\code{polr} or \code{multinom} for which (e)PRE will be calculated.}
}
\value{
An object of class \code{pre}, which is a list with the following
elements: \item{pre}{The proportional reduction in error} \item{epre}{The
expected proportional reduction in error} \item{m1form}{The formula for
model 1} \item{m2form}{The formula for model 2} \item{pcp}{The percent
correctly predicted by model 1} \item{pmc}{The percent correctly predicted
by model 2} \item{epcp}{The expected percent correctly predicted by model 1}
\item{epmc}{The expected percent correctly predicted by model 2}
}
\description{
Calculates proportional reduction in error (PRE) and expected proportional
reduction in error (epre) from Herron (1999).
}
\details{
Proportional reduction in error is calculated as a function of correct and
incorrect predictions (and the probabilities of correct and incorrect
predictions for ePRE).
}
\references{
Herron, M.  1999.  Postestimation Uncertainty in Limited
Dependent Variable Models.  Political Analysis 8(1): 83--98.
}
\author{
Dave Armstrong
}
