% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_pca.r
\name{impute_pca}
\alias{impute_pca}
\title{PCA Imputation}
\usage{
impute_pca(Y)
}
\arguments{
\item{Y}{A binary data matrix with missing data.}
}
\value{
A matrix with missing data imputed.
}
\description{
Use Principal Components to Impute Voting Data.
}
\details{
Uses the method discussed in Potthoff (2018) to impute missing data in a
binary voting data matrix.
}
