% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_latent.R
\name{init_lv}
\alias{init_lv}
\title{Initialize Latent Variable}
\usage{
init_lv(
  X,
  minprop = 0.2,
  terms,
  method = c("pca", "glrm"),
  k = 5,
  nRounds = 1,
  h2o.init.args,
  h2o.glrm.args,
  ...
)
}
\arguments{
\item{X}{A matrix of observations that are 0, 1 or \code{NA}.}

\item{minprop}{The proportion voting in the minority that serves as the
cutoff.  Votes with fewer than n times \code{minprop} votes in the minority
will be removed from the vote matrix.}

\item{terms}{A vector of values identifying the term in which the
vote was taken.}

\item{method}{A character - either "pca" or "glrm" identifying the
default method for initializing the latent variable.}

\item{k}{Scalar giving the number of dimensions to be estimated.  Note,
this does not necessarily have to be the same as the number of dimensions
estimated in the final model.}

\item{nRounds}{Number of rounds of GLRM.  If this is 1, then all bills are 
fed into the GLRM.  If ths is greater than 1, then \code{nRounds}/\code{ncol(X)} 
bills are fed into the GLRM each time and the results of the initialized latent 
variable are averaged over the different runs.  The right value will depend 
on the computational resources at hand, but we suggest the smalles value such that
\code{ncol(X)}/\code{nRounds} < 5000 as a good place to start.  If set to \code{NULL}, 
the algorithm uses \code{ceiling(ncol(X)/5000)}.}

\item{h2o.init.args}{A list of arguments to be passed to \code{h2o.init}.}

\item{h2o.glrm.args}{A list or arguments to be passed to \code{h2o.glrm}.}

\item{...}{Other arguments to be passed down - currently unimplemented.}
}
\value{
A list with the reduced set of votes, their corresponding terms
and an n x k matrix of latent variable estimates.
}
\description{
Initialize latent variable with GLRM or PCA.
}
\details{
This function initializes a static latent variable for each observation
on a given number of dimensions.  This can use a principal components
analysis (PCA) model or a generalized low-rank model (GLRM).  The latter
requires you to have the \code{h2o} package installed.
}
