% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legR.R
\name{legR}
\alias{legR}
\title{Multi-dimensional Independent Dynamic Item Response Theory}
\usage{
legR(
  X,
  terms,
  est_model = FALSE,
  legis_data = NULL,
  priors = NULL,
  k = 5,
  ndim = 2,
  nperterm = NULL,
  bestmin = 0,
  othermax = 0,
  dynIRT_control = list(threads = 1, verbose = TRUE, thresh = 1e-06, maxit = 500),
  seed = 519,
  ...
)
}
\arguments{
\item{X}{An \eqn{n\times p}{nxp} matrix of votes with values 1, 0 or \code{NA}.}

\item{terms}{A length \eqn{p}{p} vector of values identifying the term in which the
vote was taken. The first term should be coded as 1 and increase by consecutive integers.}

\item{est_model}{Logical indicating whether model should be estimated.
If \code{FALSE}, the initial latent variable is estimated, the pres
calculated and the best dimension identified.  A table is printed of the
best dimension by term.  In this case, just the PREs, best dimension
and the data with lop-sided votes removed are returned.}

\item{legis_data}{A data frame giving information about the legislators whose
votes are in \code{X}.  The only requirement is that the variable identifying the observations
be called \code{"name"}.}

\item{priors}{User specified priors to be passed to the \code{dynIRT} function.  Must be a
list of priors of the same length as the dimensionality being estimated.}

\item{k}{Number of dimensions for the GLRM to estimate.}

\item{ndim}{Number of dimensions of the dynamic IRT model to be estimated.}

\item{nperterm}{Minimum number of bills to include per term.}

\item{bestmin}{Scalar indicating the minimum PRE value to be
considered "well predicted" bythe model.}

\item{othermax}{Scalar indicating the maximum PRE of the inferior dimension.
If \code{NULL}, the algorithm just chooses the first highest PRE, regardless
of the other PRE values.}

\item{dynIRT_control}{A list containing control parameters for the dynamic
IRT model.  See the \code{\link{dynIRT}} documentation from the \code{emIRT} package.}

\item{seed}{Random number generator seed passed down to estimation functions.}

\item{...}{Other arguments to be passed down to \code{init_lv}, \code{calc_pres},
\code{remove_lop}.}
}
\value{
Depending on \code{est_model} either a list with the IRT inputs
for evaluation purposes or a list of models and a dataset of latent variable
estimates.
}
\description{
Estimates the multi-dimensional dynamic IRT model.
}
