% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Make Data for MI-DIRT}
\usage{
make_data(X, terms, legis_data = NULL, ...)
}
\arguments{
\item{X}{Matrix of votes.}

\item{terms}{Vector of term numbers corresponding to each of the columns in \code{X}.}

\item{legis_data}{A data frame minimally giving an observation identifier for the legislators.}

\item{...}{Other arguments to be passed down - currently not implemented.}
}
\value{
A list that can be passed to the \code{dynIRT()} function from \code{emIRT}.
}
\description{
Makes data to pass to the dynamic IRT function.
}
