% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_priors.R
\name{make_priors}
\alias{make_priors}
\title{Make Priors for MI-DIRT}
\usage{
make_priors(X, agree_n_thresh = 50, ...)
}
\arguments{
\item{X}{Matrix of votes.}

\item{agree_n_thresh}{Scalar indicating how many common votes legislators must have to be anchors.}

\item{...}{Other arguments to be passed down - currently not implemented.}
}
\value{
A list that can be passed to the \code{dynIRT()} function from \code{emIRT}.
}
\description{
Makes priors to pass to the dynamic IRT function.
}
