% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_starts.r
\name{make_starts}
\alias{make_starts}
\title{Make Starts for MI-DIRT}
\usage{
make_starts(data, priors, b, ...)
}
\arguments{
\item{data}{The input data to a \code{dynIRT} model.}

\item{priors}{The input priors to a \code{dynIRT} model.}

\item{b}{The coefficients from the \code{calc_pres} function.}

\item{...}{Other arguments to be passed down - currently not implemented.}
}
\value{
A list that can be passed to the \code{dynIRT()} function from \code{emIRT}.
}
\description{
Makes starting values to pass to the dynamic IRT function.
}
