% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalize.r
\name{orthogonalize}
\alias{orthogonalize}
\title{Orthogonalize Collect Data from Estimated Models}
\usage{
orthogonalize(x, ...)
}
\arguments{
\item{x}{A data frame or tibble produced with \code{gather_data}.}

\item{...}{Other arguments to be passed down, currently not implemented.}
}
\value{
A data frame with latent variable estimates.
}
\description{
Orthogonalizes the data collected using \code{gather_data} using 
the Gram-Schmidt method.
}
