% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_lop.r
\name{remove_lop}
\alias{remove_lop}
\title{Remove Lopsided Votes}
\usage{
remove_lop(X, terms, minprop = 0.2, ...)
}
\arguments{
\item{X}{A vote matrix}

\item{terms}{A vector of values identifying the term in which the
vote was taken.}

\item{minprop}{The proportion voting in the minority that serves as the
cutoff.  Votes with fewer than n times \code{minprop} votes in the minority
will be removed from the vote matrix.}

\item{...}{other arguments to be passed down - currently not implemented.}
}
\value{
A vote matrix
}
\description{
Removes lopsided votes from the voting matrix
}
